feat - Permissions Setup
Permissions Setup. Setting category which modifies data and get category which fetches data.
