+++ src/main/java/io/github/alathra/alathraskills/api/commands/TestGetExerienceCommandMemory.java
@@ -1,4 +1,4 @@
-package io.github.alathra.alathraskills.api.testing;
+package io.github.alathra.alathraskills.api.commands;
 
 import org.bukkit.entity.Player;
 
@@ -9,15 +9,16 @@ import dev.jorel.commandapi.arguments.IntegerArgument;
 import dev.jorel.commandapi.arguments.PlayerArgument;
 import dev.jorel.commandapi.executors.CommandArguments;
 import io.github.alathra.alathraskills.api.SkillsPlayerManager;
+import io.github.alathra.alathraskills.db.DatabaseQueries;
 
-public class TestAddSkillCommandMemory {
+public class TestGetExerienceCommandMemory {
 
-    public TestAddSkillCommandMemory() {
-        new CommandAPICommand("testAddSkill_memory")
-        	.withArguments(new PlayerArgument("targetPlayer"), new IntegerArgument("skill"))
-            .withFullDescription("Add Skill for a Given Player.")
-            .withShortDescription("Add Skill")
-            .withPermission("example.command")
+    public TestGetExerienceCommandMemory() {
+        new CommandAPICommand("testGetExperience_memory")
+        	.withArguments(new PlayerArgument("targetPlayer"), new IntegerArgument("skillCategoryID"))
+            .withFullDescription("Get Experience For a Given Skill Category.")
+            .withShortDescription("Get Experience")
+            .withPermission("alathraskills.get")
             .executesPlayer(this::runCommand)
             .register();
     }
@@ -30,21 +31,23 @@ public class TestAddSkillCommandMemory {
                         .build()
             );
             return;
-        }
-    	if (args.get("skill") == null) {
+        }        if (args.get("skillCategoryID") == null) {
             player.sendMessage(
-                    ColorParser.of("Provide a value after the target player to indicate the skill to add.")
+                    ColorParser.of("Provide a value after the target player to indicate skill category.")
                         .parseLegacy() // Parse legacy color codes
                         .build()
             );
+            return;
         }
-    	SkillsPlayerManager.addPlayerSkill((Player) args.get("targetPlayer"), (Integer) args.get("skill"));
+        float dbReturnValue = SkillsPlayerManager.getSkillCategoryExperience((Player) args.get("targetPlayer"), (Integer) args.get("skillCategoryID"));
         String returnString =
         		"Player with ID " +
 				((Player) args.get("targetPlayer")).getUniqueId() +
-				" has had the skill " +
-				args.get("skill") +
-				" added in memory.";
+				" has an experience value of " +
+				Float.toString(dbReturnValue) +
+				" in skill category " +
+				args.get("skillCategoryID") +
+				".";
         player.sendMessage(
                 ColorParser.of(returnString)
                     .parseLegacy() // Parse legacy color codes
