+++ src/main/java/io/github/alathra/alathraskills/api/commands/TestSetExerienceCommandMemory.java
@@ -1,4 +1,4 @@
-package io.github.alathra.alathraskills.db.testing;
+package io.github.alathra.alathraskills.api.commands;
 
 import org.bukkit.entity.Player;
 
@@ -9,16 +9,16 @@ import dev.jorel.commandapi.arguments.FloatArgument;
 import dev.jorel.commandapi.arguments.IntegerArgument;
 import dev.jorel.commandapi.arguments.PlayerArgument;
 import dev.jorel.commandapi.executors.CommandArguments;
-import io.github.alathra.alathraskills.db.DatabaseQueries;
+import io.github.alathra.alathraskills.api.SkillsPlayerManager;
 
-public class TestSetExerienceCommand {
+public class TestSetExerienceCommandMemory {
 
-    public TestSetExerienceCommand() {
-        new CommandAPICommand("testSetExperience")
+    public TestSetExerienceCommandMemory() {
+        new CommandAPICommand("testSetExperience_memory")
         	.withArguments(new PlayerArgument("targetPlayer"), new IntegerArgument("skillCategoryID"), new FloatArgument("Experience"))
-            .withFullDescription("Set Experience For a Given Skill Category.")
+            .withFullDescription("Set Experience For a Given Skill Category in Memory.")
             .withShortDescription("Set Experience")
-            .withPermission("example.command")
+            .withPermission("alathraskills.set")
             .executesPlayer(this::runCommand)
             .register();
     }
@@ -48,7 +48,7 @@ public class TestSetExerienceCommand {
             return;
         }
         // TODO Make Async
-        DatabaseQueries.saveSkillCategoryExperience(player, (Integer) args.get("skillCategoryID"), (float) args.get("Experience"));
+        SkillsPlayerManager.setPlayerExperience(player, (Integer) args.get("skillCategoryID"), (float) args.get("Experience"));
         String returnString =
         		"Player with ID " +
 				player.getUniqueId() +
