+++ src/main/java/io/github/alathra/alathraskills/command/CommandHandler.java
@@ -2,17 +2,17 @@ package io.github.alathra.alathraskills.command;
 
 import io.github.alathra.alathraskills.AlathraSkills;
 import io.github.alathra.alathraskills.Reloadable;
-import io.github.alathra.alathraskills.api.testing.TestAddSkillCommandMemory;
-import io.github.alathra.alathraskills.api.testing.TestDeleteSkillCommandMemory;
-import io.github.alathra.alathraskills.api.testing.TestGetAllSkillsCommandMemory;
-import io.github.alathra.alathraskills.api.testing.TestGetExerienceCommandMemory;
-import io.github.alathra.alathraskills.api.testing.TestSetExerienceCommandMemory;
-import io.github.alathra.alathraskills.db.testing.TestDeleteAllSkillsCommand;
-import io.github.alathra.alathraskills.db.testing.TestGetAllSkillsCommand;
-import io.github.alathra.alathraskills.db.testing.TestGetExerienceCommand;
-import io.github.alathra.alathraskills.db.testing.TestHasSkillCommand;
-import io.github.alathra.alathraskills.db.testing.TestSetExerienceCommand;
-import io.github.alathra.alathraskills.db.testing.TestSetSkillCommand;
+import io.github.alathra.alathraskills.api.commands.TestAddSkillCommandMemory;
+import io.github.alathra.alathraskills.api.commands.TestDeleteSkillCommandMemory;
+import io.github.alathra.alathraskills.api.commands.TestGetAllSkillsCommandMemory;
+import io.github.alathra.alathraskills.api.commands.TestGetExerienceCommandMemory;
+import io.github.alathra.alathraskills.api.commands.TestSetExerienceCommandMemory;
+import io.github.alathra.alathraskills.db.commands.TestDeleteAllSkillsCommand;
+import io.github.alathra.alathraskills.db.commands.TestGetAllSkillsCommand;
+import io.github.alathra.alathraskills.db.commands.TestGetExerienceCommand;
+import io.github.alathra.alathraskills.db.commands.TestHasSkillCommand;
+import io.github.alathra.alathraskills.db.commands.TestSetExerienceCommand;
+import io.github.alathra.alathraskills.db.commands.TestSetSkillCommand;
 import dev.jorel.commandapi.CommandAPI;
 import dev.jorel.commandapi.CommandAPIBukkitConfig;
