+++ src/main/java/io/github/alathra/alathraskills/db/commands/TestDeleteAllSkillsCommand.java
@@ -1,4 +1,4 @@
-package io.github.alathra.alathraskills.api.testing;
+package io.github.alathra.alathraskills.db.commands;
 
 import org.bukkit.entity.Player;
 
@@ -8,17 +8,16 @@ import dev.jorel.commandapi.CommandAPICommand;
 import dev.jorel.commandapi.arguments.IntegerArgument;
 import dev.jorel.commandapi.arguments.PlayerArgument;
 import dev.jorel.commandapi.executors.CommandArguments;
-import io.github.alathra.alathraskills.api.SkillsPlayerManager;
 import io.github.alathra.alathraskills.db.DatabaseQueries;
 
-public class TestGetExerienceCommandMemory {
+public class TestDeleteAllSkillsCommand {
 
-    public TestGetExerienceCommandMemory() {
-        new CommandAPICommand("testGetExperience_memory")
-        	.withArguments(new PlayerArgument("targetPlayer"), new IntegerArgument("skillCategoryID"))
-            .withFullDescription("Get Experience For a Given Skill Category.")
+    public TestDeleteAllSkillsCommand() {
+        new CommandAPICommand("testDeleteSkills_db")
+        	.withArguments(new PlayerArgument("targetPlayer"))
+            .withFullDescription("Deletes all skills for this player")
             .withShortDescription("Get Experience")
-            .withPermission("example.command")
+            .withPermission("alathraskills.set")
             .executesPlayer(this::runCommand)
             .register();
     }
@@ -31,23 +30,13 @@ public class TestGetExerienceCommandMemory {
                         .build()
             );
             return;
-        }        if (args.get("skillCategoryID") == null) {
-            player.sendMessage(
-                    ColorParser.of("Provide a value after the target player to indicate skill category.")
-                        .parseLegacy() // Parse legacy color codes
-                        .build()
-            );
-            return;
         }
-        float dbReturnValue = SkillsPlayerManager.getSkillCategoryExperience((Player) args.get("targetPlayer"), (Integer) args.get("skillCategoryID"));
+        // TODO Make Async
+        DatabaseQueries.deletePlayerSkills((Player) args.get("targetPlayer"));
         String returnString =
         		"Player with ID " +
 				((Player) args.get("targetPlayer")).getUniqueId() +
-				" has an experience value of " +
-				Float.toString(dbReturnValue) +
-				" in skill category " +
-				args.get("skillCategoryID") +
-				".";
+				" has had all of their skills deleted.";
         player.sendMessage(
                 ColorParser.of(returnString)
                     .parseLegacy() // Parse legacy color codes
